+++ event-management/app/Http/Controllers/Api/EventController.php
@@ -7,12 +7,19 @@ use App\Http\Resources\EventResource;
 use App\Http\Traits\CanLoadRelationships;
 use Illuminate\Http\Request;
 use App\Models\Event;
+use Illuminate\Support\Facades\Gate;
 
 class EventController extends Controller
 {
     use CanLoadRelationships;
 
     private array $relations = ['user', 'attendees', 'attendees.user']; // automatically sends relations to CanLoadRelationships
+
+    // PROTECT THE ROUTES USING USER AUTH TOKEN
+    public function __construct() 
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show']);
+    }
     /**
      * Display a listing of the resource.
      */
@@ -51,7 +58,7 @@ class EventController extends Controller
                 'start_time' =>'required|date',
                 'end_time' =>'required|date|after:start_time',
             ]),
-            'user_id' => 1
+            'user_id' => $request->user()->id,
         ]);
 
         return new EventResource($this->loadRelationships($event));
@@ -71,6 +78,11 @@ class EventController extends Controller
      */
     public function update(Request $request , Event $event)
     {
+        // if (Gate::denies('update-event', $event)) {
+        //     abort(403, 'You are not authorized to update this event.');
+        // }
+        $this->authorize('update-event', $event);
+
         $event->update($request->validate([
                 'name' =>'sometimes|string|max:255',
                 'description' =>'nullable|string',
@@ -87,6 +99,7 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
+        $this->authorize('destroy-event', $event);
         $event->delete();
 
         // return response()->json([
